Move TestData to HistoryTestsBase
